Import a DCT specification from a Google Sheets spreadsheet

Import a DCT specification from a Google Sheets spreadsheet

dct_from_gs(
  gs,
  sheets = TRUE,
  path = NULL,
  localBackup = NULL,
  preventOverwriting = psyverse::opts$get("preventOverwriting"),
  encoding = psyverse::opts$get("encoding")
)

Arguments

gs
The URL to the Google Sheets spreadsheet, its key, or the object resulting from a call to googlesheets4::gs4_get().

sheets
TRUE to read all worksheets, or a vector with the indices or names of the worksheets to read.

path
The path to save the DCT specifications.

localBackup
If pointing to a filename in an existing path, the spreadsheet will be saved as xlsx file using the openxlsx::openxlsx package.

preventOverwriting
Whether to prevent overwriting.

encoding
The encoding to use.

Value

Invisibly, an object with the worksheets and the DCT objects.