DCT templates

These functions can generate one or more empty DCT templates.

generate_dct_template(
  prefix = paste(sample(letters, 4), collapse = ""),
  output = NULL,
  overwrite = FALSE,
  createDirs = FALSE,
  addComments = TRUE,
  stopOnIllegalChars = FALSE
)

generate_dct_templates(
  x,
  outputDir = NULL,
  createDirs = FALSE,
  addComments = FALSE,
  stopOnIllegalChars = FALSE
)

Arguments

prefix, x

The prefix (prefix) or vector of prefixes (x) to use.

output, outputDir

The filename or directory to which to write the templates.

overwrite

Whether to overwrite any existing files.

createDirs

Whether to recursively create the directories if the path specified in output or outputPath does not yet exist.

addComments

Whether to add comments to the DCT specification as extra explanation.

stopOnIllegalChars

DCT identifier prefixes can only contain upper- and lowercase letters and underscores. This argument specifies whether to remove illegal characters with a warning, or whether to throw an error (and stop) if illegal characters are found,

Value

The DCT template(s), either invisibly (if output or outputDir is specified) or visibly.